Reorganize Hercules interfaces library
