+++ whoc/interfaces/__init__.py
@@ -1,7 +1,10 @@
 from whoc.interfaces.hercules_actuator_disk_interface import (
     HerculesADInterface,
-    HerculesLSADInterface,
+    HerculesHybridADInterface,
 )
 from whoc.interfaces.hercules_battery_interface import HerculesBatteryInterface
-from whoc.interfaces.hercules_hybrid_actuator_disk_interface import HerculesHybridADInterface
+from whoc.interfaces.hercules_longrun_interface import (
+    HerculesHybridLongRunInterface,
+    HerculesLongRunInterface,
+)
 from whoc.interfaces.rosco_zmq_interface import ROSCO_ZMQInterface
